+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.7] - 2026-05-26
+
+### Added
+
+- Added unit tests and benchmarks
+
+### Changed
+
+- Performance improvements on net481
+
+### Fixed
+
+- Equals fix for AttributeContextAndData and minor churn
+
 ## [1.0.6] - 2025-08-16
 
 ### Added
@@ -106,7 +120,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-[Unreleased]: https://github.com/datacute/IncrementalGeneratorExtensions/compare/1.0.6...develop
+[Unreleased]: https://github.com/datacute/IncrementalGeneratorExtensions/compare/1.0.7...develop
+[1.0.7]: https://github.com/datacute/IncrementalGeneratorExtensions/releases/tag/1.0.7
 [1.0.6]: https://github.com/datacute/IncrementalGeneratorExtensions/releases/tag/1.0.6
 [1.0.5]: https://github.com/datacute/IncrementalGeneratorExtensions/releases/tag/1.0.5
 [1.0.4]: https://github.com/datacute/IncrementalGeneratorExtensions/releases/tag/1.0.4